Moving redaction types to own files
